+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Update invalid download page
+- View only links/invites
 
 ## 2024-11-06
 - Fixed scrolling on modal closing